Moved functions for ctrl/cmd modifier key checking to separate utils file
