+++ packages/core/src/extensions/Hyperlinks/menus/helpers/PanelTextInput.tsx
@@ -1,6 +1,7 @@
 import { KeyboardEvent, PureComponent } from "react";
 import { Input } from "./PanelTextInputStyles";
 import { FocusEvent } from "react";
+import { isAppleOS } from "../../../../utils";
 
 // code adapted from https://bitbucket.org/atlassian/design-system-mirror/src/master/editor/editor-core/src/ui/PanelTextInput/index.tsx
 
@@ -29,7 +30,6 @@ export interface State {
   value?: string;
 }
 
-const isMacOS = () => /Mac/.test(navigator.platform);
 const KeyZCode = 90;
 const KeyYCode = 89;
 
@@ -141,17 +141,17 @@ export default class PanelTextInput extends PureComponent<Props, State> {
     return (
       event.keyCode === KeyZCode &&
       // cmd + z for mac
-      ((isMacOS() && event.metaKey && !event.shiftKey) ||
+      ((isAppleOS() && event.metaKey && !event.shiftKey) ||
         // ctrl + z for non-mac
-        (!isMacOS() && event.ctrlKey))
+        (!isAppleOS() && event.ctrlKey))
     );
   }
 
   private isRedoEvent(event: KeyboardEvent<any>) {
     return (
       // ctrl + y for non-mac
-      (!isMacOS() && event.ctrlKey && event.keyCode === KeyYCode) ||
-      (isMacOS() &&
+      (!isAppleOS() && event.ctrlKey && event.keyCode === KeyYCode) ||
+      (isAppleOS() &&
         event.metaKey &&
         event.shiftKey &&
         event.keyCode === KeyZCode) ||
